requirements.txt
streamlit>=1.36
pydantic>=2.7
python-dotenv>=1.0
openai>=1.40.0
mcp[cli]>=1.0.0